Refactored ClickableItem
Removed logic from ClickableItem to ClickableAudio and ClickableAnimation.
Added UnPause method of WiggleIndicator to the event that fires when the animation of that object is finished.
Renamed ClickableScriptableObject to AudioClipObject and reduced it to 1 AudioClip.
Removed EndCustomBehaviour from the interface.

(Script should decide themselves when they end or they can use the event on the ClickableAnimation script by adding themselves as a listener)