Made some adjustment to how a path is found for the zoomers.
Added zoom right variant and zoom left variant.
Added sprite for zoomer facing left at start.
Changed how the sprite is flipped (now using the SpriteRenderer).
When clicking on a zoomer which is an IClickable, a random index is chosen and used for the random list of targets given in the inspector.
Removed PathTarget.cs.